Add options to ignore items moved by hoppers and droppers.
