+++ core/src/main/java/com/griefcraft/listeners/LWCPlayerListener.java
@@ -74,95 +74,6 @@ public class LWCPlayerListener implements Listener {
         this.plugin = plugin;
     }
 
-    @EventHandler(ignoreCancelled = true)
-    public void onMoveItem(InventoryMoveItemEvent event) {
-        boolean result;
-
-        // if the initiator is the same as the source it is a dropper i.e. depositing items
-        if (event.getInitiator() == event.getSource()) {
-            result = handleMoveItemEvent(event.getInitiator(), event.getDestination());
-        } else {
-            result = handleMoveItemEvent(event.getInitiator(), event.getSource());
-        }
-
-        if (result) {
-            event.setCancelled(true);
-        }
-    }
-
-    /**
-     * Handle the item move event
-     *
-     * @param inventory
-     */
-    private boolean handleMoveItemEvent(Inventory initiator, Inventory inventory) {
-        LWC lwc = LWC.getInstance();
-
-        if (inventory == null) {
-            return false;
-        }
-
-        Location location;
-        InventoryHolder holder;
-        Location hopperLocation = null;
-        InventoryHolder hopperHolder;
-
-        try {
-            holder = inventory.getHolder();
-            hopperHolder = initiator.getHolder();
-        } catch (AbstractMethodError e) {
-            return false;
-        }
-
-        try {
-            if (holder instanceof BlockState) {
-                location = ((BlockState) holder).getLocation();
-            } else if (holder instanceof DoubleChest) {
-                location = ((DoubleChest) holder).getLocation();
-            } else {
-                return false;
-            }
-
-            if (hopperHolder instanceof Hopper) {
-                hopperLocation = ((Hopper) hopperHolder).getLocation();
-            }
-        } catch (Exception e) {
-            return false;
-        }
-
-        // High-intensity zone: increase protection cache if it's full, otherwise
-        // the database will be getting rammed
-        lwc.getProtectionCache().increaseIfNecessary();
-
-        // Attempt to load the protection at that location
-        Protection protection = lwc.findProtection(location);
-
-        // If no protection was found we can safely ignore it
-        if (protection == null) {
-            return false;
-        }
-
-        if (hopperLocation != null && Boolean.parseBoolean(lwc.resolveProtectionConfiguration(Material.HOPPER, "enabled"))) {
-            Protection hopperProtection = lwc.findProtection(hopperLocation);
-
-            if (hopperProtection != null) {
-                // if they're owned by the same person then we can allow the move
-                if (hopperProtection.getOwner().equals(protection.getOwner())) {
-                    return false;
-                }
-            }
-        }
-
-        boolean denyHoppers = Boolean.parseBoolean(lwc.resolveProtectionConfiguration(Material.getMaterial(protection.getBlockId()), "denyHoppers"));
-
-        // xor = (a && !b) || (!a && b)
-        if (denyHoppers ^ protection.hasFlag(Flag.Type.HOPPER)) {
-            return true;
-        }
-
-        return false;
-    }
-
     @EventHandler(priority = EventPriority.HIGH)
     public void onPlayerDropItem(PlayerDropItemEvent event) {
         if (event.isCancelled() || !LWC.ENABLED) {
